new ANSI C implementation of HTM with emphasis on better performance
New ANSI C implementation of HTM.  The initial commit is a baseline
single-threaded version. Next I hope to expand upon this with
multithread support.  Longer term this will be the implementation that
is very performance focused, trying to get HTM to run as fast and lean
as we can on larger and larger data sets.